Added authentication and authorization

+++ app/directors/routes.py
@@ -1,4 +1,5 @@
 from flask import render_template, request, Blueprint, jsonify
+from flask_login import login_user, current_user, logout_user, login_required
 from flask_sqlalchemy import SQLAlchemy
 from ..models import User, Movie, Director, Genre
 from .. import db
@@ -17,29 +18,36 @@ def directors():
         else:
             return 'Nothing found', 404
     elif request.method == 'POST':
-        new_name = request.form['name']
-        new_director = Director(name=new_name)
-        db.session.add(new_director)
-        db.session.flush()
-        db.session.commit()
+        if current_user.is_admin:
+            new_name = request.form['name']
+            new_director = Director(name=new_name)
+            db.session.add(new_director)
+            db.session.flush()
+            db.session.commit()
 
-        return jsonify(new_director.to_json()), 201
+            return jsonify(new_director.to_json()), 201
+        else:
+            return 'You do not have permission for this!', 403
 
 
 @directors_bp.route('/director/<int:id>', methods=['PUT', 'DELETE'])
+@login_required
 def director(id):
-    if request.method == 'PUT':
-        director = Director.get_or_404(id)
-        new_name = request.form['name']
-        director.name = new_name
-        db.session.flush()
-        db.session.commit()
+    if current_user.is_admin:
+        if request.method == 'PUT':
+            director = Director.get_or_404(id)
+            new_name = request.form['name']
+            director.name = new_name
+            db.session.flush()
+            db.session.commit()
 
-        return jsonify(director), 201
-    elif request.method == 'DELETE':
-        director = Director.query.filter_by(director=id).delete()
+            return jsonify(director), 201
+        elif request.method == 'DELETE':
+            director = Director.query.filter_by(director=id).delete()
 
-        if not director:
-            return 'Director with that id not found', 404
+            if not director:
+                return 'Director with that id not found', 404
 
-        return 'Deleted!', 201
+            return 'Deleted!', 201
+    else:
+        return 'You do not have permission for this!', 403
